feat(core): replace hook system with ERC20VotesComp
The hook system was designed to add flexibility to the ERC20 so that in
the future a snaspshotting mechanism like the one in the COMP token
could be "hooked".

Luckily for us, the OpenZeppelin team developed ERC20VotesComp, with
provides exactly this functionality, so we can import managed and tested
code into our contract, removing the need for the hook system.

ERC20VotesComp is used instead of ERC20Votes to maintain compatibility
with Coumpound's GovernorBravo. In OpenZeppelin 4.3 another governor
system is supposedly coming, but this opens up the door to use
GovernorBravo right now, and the max supply limitations are not a
concern for us.

See: https://blog.openzeppelin.com/openzeppelin-contracts-4-2/