refactor(core): remove GovernorBravo upgradeability
The main reason I'm removing upgradeability is because I hit a bug that I
couldn't debug: the proxy pattern delegation was working for some
methods and failing for others. For example, calling _setPendingAdmin of
GovernorBravoDelegate through GovernorBravoDelegator worked, but calling
cancel or propose didn't, and I couldn't figure out why.

Besides, the reason I kept upgradeability in the first place is because
I didn't want to modify Compound's code more than necessary, but I was
never a fan of it:

It complicates the deployment process and adds friction to everything:
making calls from web3 isn't as simple, the truffle debugger doesn't
(appear to) work out of the box with delegatecall, etc.

And the most important reason: I don't believe we'll ever need upgradeability.
Thanks to the Governor-Timelock separation (a design that I like a lot) in the
event that we ever needed to migrate the Governor system we could deploy a
"GovernorCharlie", make Timelock admin of it, and make GovernorBravo renounce
ownership of Timelock in favour of GovernorCharlie, thus migrating governacne
systems. Sure, we wouldn't inherit the old state seamlessly, but... is that so
bad?

Personally I believe that upgradeability is not worth the tradeoffs and
prefer a simpler system that is easier to understand.