feat(core): simplify MintControl
The rationale behind MintControl was to give Dandercoin a certain
safeguard mechanism while it was kept centralized. Thus, the production
of new tokens would be centralized and could be delegated to different
addresses.

However, now we can ship with governance, which makes the feature of
minter authorization unnecessary. Instead of authorizing one minter to
mint 10000 DANDER per year, have governance give them 10000 DANDER; it
is essentially the same thing.

Plus, "big mints", such as when an employee joins the company and is
entitled to 1000 DANDER should be voted by governance anyway.

This commit removes the ability to authorize individual minters and
keeps the global limit to retain the protocol level safety net. Less
code = less liability.